+++ CHANGELOG.md
@@ -6,6 +6,17 @@ All notable changes to the "**MinifyAll**" VS Code extension will be documented 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ---
+## [**Current**]
+
+### Changed
+
+* Most settings will now be updated silently in the background and not require a manual reload. Some special settings require a reload, in which case
+  the user is prompted for it. (as discussed in Issue #132)
+
+### Removed
+
+* The global variable for the `terser` options is removed. The settings are used directly so they actually update.
+* The global variable `minifyHex` is removed, as it is only ever used once. Now the setting is used directly.
 
 ## [**2.9.4**] - 2021-10-14